Add files via upload
We design two pages one is the product page with a search bar at the top then the category list at the end list of products. And another one is the details page provides you more details about the product like price, long description, at the bottom chat, and add to cart button.